refactor endpoints into mock_server module
