+++ lib/davinci_pdex_test_kit/mock_server/resources/mock_capability_statement.json
